third_party/blink/web_tests/external/wpt/editing/run/inserttext_1001-2000-expected.txt

This is a testharness.js-based test.
Found 7 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["inserttext","\\t"]] "http://a[]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "http://a\\t" but got "http://a<span style=\\"white-space:pre\\">\\t</span>"
[FAIL] [["inserttext","a"]] "<a href=/>foo[]</a>bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<a href=\\"/\\">fooa</a>bar" but got "<a href=\\"/\\">foo</a>abar"
[FAIL] [["defaultparagraphseparator","div"],["inserttext","a"]] "<p>fo[o<p>b]ar" queryCommandValue("defaultparagraphseparator") before
  assert_equals: Wrong result returned expected "p" but got "div"
[FAIL] [["stylewithcss","true"],["inserttext","a"]] "<p>[foo<span style=color:#aBcDeF>bar]</span>baz" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["inserttext","a"]] "<p>foo<span style=color:#aBcDeF>[bar</span>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo<span style=\\"color:rgb(171, 205, 239)\\">a</span></p>" but got "<p>foo<font color=\\"#abcdef\\">a</font></p>"
[FAIL] [["inserttext","a"]] "<p>foo<span style=color:#aBcDeF>{bar</span>baz}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo<span style=\\"color:rgb(171, 205, 239)\\">a</span></p>" but got "<p>foo<font color=\\"#abcdef\\">a</font></p>"
[FAIL] [["stylewithcss","false"],["inserttext","a"]] "<p>foo<span style=color:#aBcDeF>[bar</span><span style=color:#fEdCbA>baz]</span>quz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo<span style=\\"color:rgb(171, 205, 239)\\">a</span>quz</p>" but got "<p>foo<font color=\\"#abcdef\\">a</font>quz</p>"
Harness: the test ran to completion.
